main structure types file instead of having like 4 files for types
