+++ src/commands/8ball.ts
@@ -2,7 +2,7 @@ import FluorineClient from '@classes/Client';
 import Embed from '@classes/Embed';
 import { CommandInteraction } from 'discord.js';
 import { SlashCommandBuilder } from '@discordjs/builders';
-import { Category } from 'types/applicationCommand';
+import { Category } from 'types/structures';
 import hash from 'murmurhash-v3';
 
 export async function run(client: FluorineClient, interaction: CommandInteraction) {
