+++ src/commands/help.ts
@@ -2,7 +2,7 @@ import FluorineClient from '@classes/Client';
 import Embed from '@classes/Embed';
 import { CommandInteraction, EmbedFieldData, MessageActionRow, MessageSelectMenu } from 'discord.js';
 import { SlashCommandBuilder } from '@discordjs/builders';
-import { Category, ChatInputCommand } from 'types/applicationCommand';
+import { Category, ChatInputCommand } from 'types/structures';
 
 export async function run(client: FluorineClient, interaction: CommandInteraction) {
     const category = interaction.options.getString('category');
