+++ src/commands/rob.ts
@@ -1,7 +1,7 @@
 import FluorineClient from '@classes/Client';
 import { CommandInteraction } from 'discord.js';
 import { SlashCommandBuilder } from '@discordjs/builders';
-import { Category } from 'types/applicationCommand';
+import { Category } from 'types/structures';
 import Embed from '@classes/Embed';
 export async function run(client: FluorineClient, interaction: CommandInteraction) {
     const user = interaction.options.getUser('user');
