+++ src/commands/withdraw.ts
@@ -1,7 +1,7 @@
 import FluorineClient from '@classes/Client';
 import { SlashCommandBuilder } from '@discordjs/builders';
 import { CommandInteraction } from 'discord.js';
-import { Category } from 'types/applicationCommand';
+import { Category } from 'types/structures';
 
 export async function run(client: FluorineClient, interaction: CommandInteraction) {
     const toWithdraw = interaction.options.getInteger('amount');
