--- src/types/event.ts
@@ -1,5 +1,0 @@
-import FluorineClient from '@classes/Client';
-
-export interface Event {
-    run: (client: FluorineClient, ...args) => void;
-}
